requirements.txt
flask
python-dotenv>=1.0
pydantic>=2
requests
pillow>=10
beautifulsoup4
pgeocode
playwright
playwright-stealth
thefuzz
sentence-transformers>=2.2
# Postgres + pgvector (semantic search, master catalog, BMW OEM embeddings)
psycopg[binary]>=3.2
pgvector>=0.3
bcrypt>=4.1
ollama>=0.3
crawl4ai>=0.8,<0.9

# Browser binaries (not installed by pip). After pip install, run ONE of:
#   ./scripts/install_scraper_browsers.sh
# Or two lines, with nothing after the browser name on each line:
#   python3 -m playwright install chromium
#   python3 -m patchright install chromium
# Do not append comments on the same line as "playwright install" unless your shell treats # as a comment; otherwise Playwright treats extra words as invalid browser names.

# Optional: encrypted users.db (set USERS_DB_ENCRYPTION_KEY). May require SQLCipher libs on some platforms.
# sqlcipher3>=0.5.4
